+++ src/Entities/PollAnswer.php
@@ -1,0 +1,36 @@
+<?php
+
+/**
+ * This file is part of the TelegramBot package.
+ *
+ * (c) Avtandil Kikabidze aka LONGMAN <akalongman@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Longman\TelegramBot\Entities;
+
+/**
+ * Class PollAnswer
+ *
+ * This entity represents an answer of a user in a non-anonymous poll.
+ *
+ * @link https://core.telegram.org/bots/api#pollanswer
+ *
+ * @method string getPollId()    Unique poll identifier
+ * @method User   getUser()      The user, who changed the answer to the poll
+ * @method array  getOptionIds() 0-based identifiers of answer options, chosen by the user. May be empty if the user retracted their vote.
+ */
+class PollAnswer extends Entity
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected function subEntities()
+    {
+        return [
+            'user' => User::class,
+        ];
+    }
+}