+++ src/Request.php
@@ -12,10 +12,12 @@
 namespace Longman\TelegramBot;
 
 use GuzzleHttp\Client;
+use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Psr7\Stream;
 use Longman\TelegramBot\Entities\File;
 use Longman\TelegramBot\Entities\InputMedia\InputMedia;
+use Longman\TelegramBot\Entities\Message;
 use Longman\TelegramBot\Entities\ServerResponse;
 use Longman\TelegramBot\Exception\InvalidBotTokenException;
 use Longman\TelegramBot\Exception\TelegramException;
@@ -54,7 +56,7 @@ use Longman\TelegramBot\Exception\TelegramException;
  * @method static ServerResponse promoteChatMember(array $data)               Use this method to promote or demote a user in a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Pass False for all boolean parameters to demote a user. Returns True on success.
  * @method static ServerResponse setChatAdministratorCustomTitle(array $data) Use this method to set a custom title for an administrator in a supergroup promoted by the bot. Returns True on success.
  * @method static ServerResponse setChatPermissions(array $data)              Use this method to set default chat permissions for all members. The bot must be an administrator in the group or a supergroup for this to work and must have the can_restrict_members admin rights. Returns True on success.
- * @method static ServerResponse exportChatInviteLink(array $data)            Use this method to generate a new invite link for a chat; any previously generated link is revoked. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns the new invite link as String on success.
+ * @method static ServerResponse exportChatInviteLink(array $data)            Use this method to generate a new invite link for a chat. Any previously generated link is revoked. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns the new invite link as String on success.
  * @method static ServerResponse setChatPhoto(array $data)                    Use this method to set a new profile photo for the chat. Photos can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
  * @method static ServerResponse deleteChatPhoto(array $data)                 Use this method to delete a chat photo. Photos can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
  * @method static ServerResponse setChatTitle(array $data)                    Use this method to change the title of a chat. Titles can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns True on success.
@@ -112,16 +114,9 @@ class Request
     /**
      * Guzzle Client object
      *
-     * @var Client
+     * @var ClientInterface
      */
     private static $client;
-
-    /**
-     * Input value of the request
-     *
-     * @var string
-     */
-    private static $input;
 
     /**
      * Request limiter
@@ -269,15 +264,9 @@ class Request
      * Initialize
      *
      * @param Telegram $telegram
-     *
-     * @throws TelegramException
      */
     public static function initialize(Telegram $telegram)
     {
-        if (!($telegram instanceof Telegram)) {
-            throw new TelegramException('Invalid Telegram pointer!');
-        }
-
         self::$telegram = $telegram;
         self::setClient(self::$client ?: new Client(['base_uri' => self::$api_base_uri]));
     }
@@ -285,16 +274,10 @@ class Request
     /**
      * Set a custom Guzzle HTTP Client object
      *
-     * @param Client $client
-     *
-     * @throws TelegramException
+     * @param ClientInterface $client
      */
-    public static function setClient(Client $client)
+    public static function setClient(ClientInterface $client)
     {
-        if (!($client instanceof Client)) {
-            throw new TelegramException('Invalid GuzzleHttp\Client pointer!');
-        }
-
         self::$client = $client;
     }
 
@@ -307,7 +290,8 @@ class Request
     public static function getInput()
     {
         // First check if a custom input has been set, else get the PHP input.
-        if (!($input = self::$telegram->getCustomInput())) {
+        $input = self::$telegram->getCustomInput();
+        if (empty($input)) {
             $input = file_get_contents('php://input');
         }
 
@@ -316,11 +300,9 @@ class Request
             throw new TelegramException('Input must be a string!');
         }
 
-        self::$input = $input;
+        TelegramLog::update($input);
 
-        TelegramLog::update(self::$input);
-
-        return self::$input;
+        return $input;
     }
 
     /**
@@ -627,6 +609,12 @@ class Request
 
         if (!$response->isOk() && $response->getErrorCode() === 401 && $response->getDescription() === 'Unauthorized') {
             throw new InvalidBotTokenException();
+        }
+
+        // Special case for sent polls, which need to be saved specially.
+        // @todo Take into account if DB gets extracted into separate module.
+        if ($response->isOk() && ($message = $response->getResult()) && ($message instanceof Message) && $poll = $message->getPoll()) {
+            DB::insertPollRequest($poll);
         }
 
         // Reset current action after completion.